web_mongo_app/pyproject.toml
[project]
name = "web-mongo-app"
version = "0.1.0"
description = "Web application uses MongoDB and FastAPI"
authors = ["Piotr Lipinski <kinetics20@gmail.com>"]
license = "Beerware"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "motor>=3.7.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
